requirements.txt
flask>=2.0.1
numpy>=1.20.0
opencv-python>=4.5.0
scikit-image>=0.18.0
pillow>=8.0.0
matplotlib>=3.4.0 